# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cognito Bughunt Teste",
    "description": "Documentação Cognito Bughunt Teste",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Healthy",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/get-idp-url": {
      "post": {
        "summary": "Pegar a URL de um usuário corporativo",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Email Corporativo.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "joelton@caiotelles.com.br"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/auth/validate-token": {
      "post": {
        "summary": "Validar o Token do Usuário corporativo.",
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/auth/list-applications": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "deprecated": true
      }
    }
  }
}